scripts/requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==4.2.0
argcomplete==3.2.1
asgiref==3.7.2
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.23
click==8.1.7
coloredlogs==15.0.1
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.7.0
distro-info==1.1+ubuntu0.1
docker==7.0.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.1
fire==0.5.0
FLAML==2.1.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gast==0.5.4
google-ai-generativelanguage==0.4.0
google-api-core==2.16.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-cloud-vision==3.6.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httplib2==0.20.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==2.8
importlib-metadata==6.11.0
importlib_resources==6.1.2
ipython==8.22.1
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
keras==2.15.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.5
kubernetes==29.0.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
linkedin==0.1.5
linkedin-api-client==0.3.0
linkedin-python==0.0.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mmh3==4.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
numpy==1.26.2
oauth2==1.9.0.post1
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.8.0
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
opt-einsum==3.3.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandas-stubs==2.1.4.231227
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.0
platformdirs==4.1.0
posthog==3.4.2
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.23.4
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyautogen @ git+https://github.com/microsoft/autogen.git@5e50def5e52f6e5df3b55cd0a19f121cb39e30af
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.14.6
pydash==7.0.7
PyGithub==2.2.0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==4.0.2
PyPika==0.48.9
pyproject_hooks==1.0.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
systemd-python==234
tenacity==8.2.3
tensorboard==1.12.2
tensorboard-data-server==0.7.2
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.12.0-py3-none-any.whl
tensorflow-estimator==1.14.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
tiktoken==0.5.2
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.2
traitlets==5.14.1
typer==0.9.0
types-pytz==2023.3.1.1
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
userpath==1.9.1
uvicorn==0.27.1
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
zipp==1.0.0
